Simplify qirlib module structure
